Reorganize order components in customer module.
