+++ client/src/app/modules/auth/modules/customer/customer-routing.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { CustomerLayoutComponent } from './components/customer-layout/customer-layout.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { OrdersComponent } from './components/orders/orders.component';
-import { NewOrderComponent } from './components/new-order/new-order.component';
-import { OrderDetailComponent } from './components/order-detail/order-detail.component';
+import { OrdersComponent } from './components/orders/orders/orders.component';
+import { NewOrderComponent } from './components/orders/new-order/new-order.component';
+import { OrderDetailComponent } from './components/orders/order-detail/order-detail.component';
 import { StatementsComponent } from './components/statements/statements/statements.component';
 import { StatementDetailComponent } from './components/statements/statement-detail/statement-detail.component';
 import { DataResolver } from './services/data.resolver';
